requirements.txt
chromadb
loguru
numpy
open_clip_torch
opencv-python
pandas
pillow
plotly
pydantic
streamlit
transformers
watchfiles
